// app/backend/src/interfaces/ITeam.ts
export default interface ITeam {
  id?: number;
  teamName: string;
}

export interface ILeaderboardTeam {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: number;
}